refactor(network): split tcp.rs and tls_tcp.rs into module folders
tcp.rs (473 lines) -> transport/tcp/
  mod.rs       -- re-exports
  transport.rs -- PeerConn + TcpTransport + basic TCP sync test
  peer.rs      -- TcpPeerTransport + concurrent hub test

tls_tcp.rs (394 lines) -> transport/tls_tcp/
  mod.rs       -- re-exports
  transport.rs -- TlsStream + TlsPeerConn + TlsTcpTransport + TLS tests
  peer.rs      -- TlsPeerTransport

Cleanup included:
- PeerConn::new() / TlsPeerConn::new() constructors eliminate 3x
  repeated struct literal construction
- raw_accept() / make_server_stream() helper methods eliminate
  repeated set_nonblocking(false/true) and ServerConnection::new()
  patterns across accept_peer / accept_any / accept_split
- TcpPeerTransport::new() / TlsPeerTransport::new() are pub(super)
  so construction is encapsulated in the module boundary
