+++ crates/zamsync-network/src/transport/tls_tcp/transport.rs
@@ -11,15 +11,19 @@ use tracing::{info, warn};
 use zamsync_core::ports::Transport;
 use zamsync_core::{NodeId, SyncMessage, ZamError, ZamResult};
 
-/// Unified read+write wrapper over a TLS stream, which can be either
-/// server-side (accepted) or client-side (connected).
-enum TlsStream {
+use super::peer::TlsPeerTransport;
+
+// ---- TLS stream wrapper -----------------------------------------------------
+
+/// Unified read+write wrapper over a TLS stream (server-accepted or client-connected).
+/// `pub(super)` so `peer.rs` can use it as a field type without exposing it publicly.
+pub(super) enum TlsStream {
     Server(StreamOwned<ServerConnection, TcpStream>),
     Client(StreamOwned<ClientConnection, TcpStream>),
 }
 
 impl TlsStream {
-    fn set_read_timeout(&self, dur: Option<Duration>) -> io::Result<()> {
+    pub(super) fn set_read_timeout(&self, dur: Option<Duration>) -> io::Result<()> {
         match self {
             TlsStream::Server(s) => s.get_ref().set_read_timeout(dur),
             TlsStream::Client(s) => s.get_ref().set_read_timeout(dur),
@@ -52,13 +56,29 @@ impl Write for TlsStream {
     }
 }
 
+// ---- Internal per-connection state ------------------------------------------
+
 struct TlsPeerConn {
     stream: TlsStream,
-    /// Bytes accumulated across 50ms poll cycles (same fix as TcpTransport).
-    frame_buf: crate::protocol::FrameBuffer,
-    /// One message buffered during accept (the initial Handshake).
+    frame_buf: protocol::FrameBuffer,
     pending: Option<SyncMessage>,
 }
+
+impl TlsPeerConn {
+    fn new(stream: TlsStream, pending: Option<SyncMessage>) -> Self {
+        Self {
+            stream,
+            frame_buf: protocol::FrameBuffer::new(),
+            pending,
+        }
+    }
+}
+
+// ---- TlsTcpTransport --------------------------------------------------------
+
+/// Fixed SNI hostname used in all ZamSync TLS connections.
+/// Must match the SAN in the node certificate generated by `zamsync keygen`.
+const TLS_SERVER_NAME: &str = "zamsync.local";
 
 /// TLS-encrypted TCP transport using mutual TLS (mTLS).
 ///
@@ -71,10 +91,6 @@ pub struct TlsTcpTransport {
     client_config: Arc<rustls::ClientConfig>,
     peers: HashMap<u32, TlsPeerConn>,
 }
-
-/// Fixed SNI hostname used in all ZamSync TLS connections.
-/// Must match the SAN in the node certificate generated by `zamsync keygen`.
-const TLS_SERVER_NAME: &str = "zamsync.local";
 
 impl TlsTcpTransport {
     pub fn bind(addr: &str, config: &TlsConfig) -> ZamResult<Self> {
@@ -96,20 +112,26 @@ impl TlsTcpTransport {
         Ok(self.listener.local_addr()?)
     }
 
+    fn raw_accept(&mut self) -> ZamResult<(TcpStream, SocketAddr)> {
+        self.listener.set_nonblocking(false)?;
+        let result = self.listener.accept()?;
+        self.listener.set_nonblocking(true)?;
+        Ok(result)
+    }
+
+    fn make_server_stream(&self, tcp: TcpStream) -> ZamResult<TlsStream> {
+        let conn = ServerConnection::new(Arc::clone(&self.server_config))
+            .map_err(|e| ZamError::Config(format!("TLS server init: {e}")))?;
+        Ok(TlsStream::Server(StreamOwned::new(conn, tcp)))
+    }
+
     /// Blocking accept: reads the initial Handshake to discover peer NodeId.
     /// Performs TLS handshake + mTLS client-cert verification transparently.
     pub fn accept_any(&mut self) -> ZamResult<NodeId> {
-        self.listener.set_nonblocking(false)?;
-        let (tcp, addr) = self.listener.accept()?;
-        self.listener.set_nonblocking(true)?;
-
-        // Give enough time for both TLS and the application Handshake.
+        let (tcp, addr) = self.raw_accept()?;
         tcp.set_read_timeout(Some(Duration::from_millis(5_000)))?;
 
-        let conn = ServerConnection::new(Arc::clone(&self.server_config))
-            .map_err(|e| ZamError::Config(format!("TLS server init: {e}")))?;
-        let mut tls = TlsStream::Server(StreamOwned::new(conn, tcp));
-
+        let mut tls = self.make_server_stream(tcp)?;
         let msg = protocol::decode(&mut tls)?;
         let node_id = match &msg {
             SyncMessage::Handshake { node_id, .. } => *node_id,
@@ -122,14 +144,7 @@ impl TlsTcpTransport {
         };
 
         tls.set_read_timeout(Some(Duration::from_millis(50)))?;
-        self.peers.insert(
-            node_id.0,
-            TlsPeerConn {
-                stream: tls,
-                frame_buf: crate::protocol::FrameBuffer::new(),
-                pending: Some(msg),
-            },
-        );
+        self.peers.insert(node_id.0, TlsPeerConn::new(tls, Some(msg)));
         info!(peer = node_id.0, %addr, "TLS peer accepted");
         Ok(node_id)
     }
@@ -151,14 +166,7 @@ impl TlsTcpTransport {
             .map_err(|e| ZamError::Config(format!("TLS client init: {e}")))?;
 
         let tls = TlsStream::Client(StreamOwned::new(conn, tcp));
-        self.peers.insert(
-            peer_id.0,
-            TlsPeerConn {
-                stream: tls,
-                frame_buf: crate::protocol::FrameBuffer::new(),
-                pending: None,
-            },
-        );
+        self.peers.insert(peer_id.0, TlsPeerConn::new(tls, None));
         info!(peer = peer_id.0, addr, "TLS connection established");
         Ok(())
     }
@@ -173,15 +181,10 @@ impl TlsTcpTransport {
     /// HashMap. The returned [`TlsPeerTransport`] is `Send` and can be moved
     /// into a worker thread for concurrent hub serving.
     pub fn accept_split(&mut self) -> ZamResult<TlsPeerTransport> {
-        self.listener.set_nonblocking(false)?;
-        let (tcp, addr) = self.listener.accept()?;
-        self.listener.set_nonblocking(true)?;
+        let (tcp, addr) = self.raw_accept()?;
         tcp.set_read_timeout(Some(Duration::from_millis(5_000)))?;
 
-        let conn = ServerConnection::new(Arc::clone(&self.server_config))
-            .map_err(|e| ZamError::Config(format!("TLS server init: {e}")))?;
-        let mut tls = TlsStream::Server(StreamOwned::new(conn, tcp));
-
+        let mut tls = self.make_server_stream(tcp)?;
         let msg = protocol::decode(&mut tls)?;
         let node_id = match &msg {
             SyncMessage::Handshake { node_id, .. } => *node_id,
@@ -195,12 +198,7 @@ impl TlsTcpTransport {
 
         tls.set_read_timeout(Some(Duration::from_millis(50)))?;
         info!(peer = node_id.0, %addr, "TLS peer accepted (split mode)");
-        Ok(TlsPeerTransport {
-            peer_id: node_id,
-            stream: tls,
-            frame_buf: crate::protocol::FrameBuffer::new(),
-            pending: Some(msg),
-        })
+        Ok(TlsPeerTransport::new(node_id, tls, Some(msg)))
     }
 }
 
@@ -232,47 +230,6 @@ impl Transport for TlsTcpTransport {
             }
         }
         Ok(None)
-    }
-}
-
-/// A single-connection TLS transport returned by [`TlsTcpTransport::accept_split`].
-///
-/// Owns exactly one TLS stream and implements [`Transport`] for that peer.
-/// `Send`-safe (both rustls `StreamOwned` variants are `Send`): move it into
-/// a worker thread so the hub can serve N TLS peers concurrently.
-pub struct TlsPeerTransport {
-    peer_id: NodeId,
-    stream: TlsStream,
-    frame_buf: protocol::FrameBuffer,
-    pending: Option<SyncMessage>,
-}
-
-impl TlsPeerTransport {
-    /// NodeId extracted from the peer's opening Handshake.
-    pub fn peer_id(&self) -> NodeId {
-        self.peer_id
-    }
-}
-
-impl Transport for TlsPeerTransport {
-    fn send(&mut self, _peer_id: NodeId, message: &SyncMessage) -> ZamResult<()> {
-        let mut writer = BufWriter::new(&mut self.stream);
-        protocol::encode(message, &mut writer)
-    }
-
-    fn receive(&mut self) -> ZamResult<Option<(NodeId, SyncMessage)>> {
-        if let Some(msg) = self.pending.take() {
-            return Ok(Some((self.peer_id, msg)));
-        }
-        match self.frame_buf.try_read_frame(&mut self.stream) {
-            Ok(Some(bytes)) => {
-                let msg = rkyv::from_bytes::<SyncMessage>(&bytes)
-                    .map_err(|e| ZamError::Serialization(format!("{}", e)))?;
-                Ok(Some((self.peer_id, msg)))
-            }
-            Ok(None) => Ok(None),
-            Err(e) => Err(e),
-        }
     }
 }
 
@@ -314,7 +271,6 @@ mod tests {
             };
             client.send(server_id, &hs).unwrap();
 
-            // Wait for SyncComplete from server
             loop {
                 if let Some((_id, msg)) = client.receive().unwrap() {
                     return msg;
@@ -323,11 +279,9 @@ mod tests {
             }
         });
 
-        // Server: accept and echo SyncComplete
         let peer_id = server.accept_any().unwrap();
         assert_eq!(peer_id.0, 1);
 
-        // The Handshake was buffered -- receive it
         let (from, msg) = loop {
             if let Some(pair) = server.receive().unwrap() {
                 break pair;
@@ -336,7 +290,6 @@ mod tests {
         assert_eq!(from.0, 1);
         assert!(matches!(msg, SyncMessage::Handshake { .. }));
 
-        // Reply with SyncComplete
         server.send(peer_id, &SyncMessage::SyncComplete).unwrap();
 
         let reply = client_handle.join().unwrap();
@@ -347,7 +300,6 @@ mod tests {
     fn test_tls_rejects_untrusted_client() {
         crate::tls::install_crypto_provider();
 
-        // Server uses creds from deployment A
         let creds_a = generate_credentials().expect("keygen A");
         let server_config = TlsConfig::from_pem(
             creds_a.node_cert_pem.clone(),
@@ -355,12 +307,11 @@ mod tests {
             creds_a.ca_cert_pem.clone(),
         );
 
-        // Client uses creds from a different deployment (B) -- should be rejected
         let creds_b = generate_credentials().expect("keygen B");
         let client_config = TlsConfig::from_pem(
             creds_b.node_cert_pem,
             creds_b.node_key_pem,
-            creds_a.ca_cert_pem.clone(), // client trusts server's CA but presents wrong cert
+            creds_a.ca_cert_pem.clone(), // trusts server CA but presents wrong cert
         );
 
         let mut server = TlsTcpTransport::bind("127.0.0.1:0", &server_config).unwrap();
@@ -371,7 +322,6 @@ mod tests {
             let server_id = NodeId(99);
             client.connect(server_id, &server_addr).unwrap();
 
-            // Try to send -- TLS handshake happens here; server rejects client cert
             let hs = SyncMessage::Handshake {
                 node_id: NodeId(2),
                 vv: VersionVector::new(),
@@ -379,12 +329,9 @@ mod tests {
             client.send(server_id, &hs)
         });
 
-        // Server-side: accept_any should fail (TLS handshake rejected)
         let server_result = server.accept_any();
-
         let client_result = client_handle.join().unwrap();
 
-        // At least one side must observe an error
         assert!(
             server_result.is_err() || client_result.is_err(),
             "expected TLS rejection but both sides succeeded"
